% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{can_compile}
\alias{can_compile}
\title{Test if compilation is possible}
\usage{
can_compile()
}
\value{
A logical scalar
}
\description{
Test if compilation appears possible.  This is used in some
examples, and tries compiling a trivial C program with \code{R CMD
SHLIB}.  Results are cached between runs within a session so this
should be fast to rely on.
}
\examples{
can_compile() # will take ~0.1s the first time
can_compile() # should be basically instantaneous
}

