% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\docType{data}
\name{can_compile_result}
\alias{can_compile_result}
\title{Test if compilation is possible}
\format{An object of class \code{NULL} of length 0.}
\usage{
can_compile_result
}
\value{
A logical scalar
}
\description{
Test if compilation appears possible.  This is used in some
examples, and tries compiling a trivial C program with \code{R CMD
SHLIB}.  Results are cached between runs within a session so this
should be fast to rely on.
}
\examples{
can_compile()
}
\keyword{datasets}

