% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{interpolation_function}
\alias{interpolation_function}
\title{Interpolation}
\usage{
interpolation_function(x, y, type)
}
\arguments{
\item{x}{Independent variable (time in an actual odin model)}

\item{y}{Dependent variable (vector or matrix)}

\item{type}{Character string indicating the interpolation type
("constant", "linear" or "spline").}
}
\description{
Interface to the same interpolation functions as odin uses.  This
is primarily going to be useful for testing so that you can see
exactly how odin interpolates functions.
}
