% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{ode_system_generator}
\alias{ode_system_generator}
\title{Generate ODE system}
\usage{
ode_system_generator(dll, name = NULL)
}
\arguments{
\item{dll}{Name of the shared library (with no extension)}

\item{name}{Name of the model within the shared library}
}
\description{
Generate an ODE interface from an odin shared library.  This is
used internally when generating packages, and should not be used
by end users.
}
\keyword{internal}

