% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_bundle.R
\name{odin_js_example}
\alias{odin_js_example}
\title{Generate a built in example}
\usage{
odin_js_example(filename, dest = tempfile())
}
\arguments{
\item{filename}{Filename of model to include}

\item{dest}{Destination directory - multiple files will be created
here, overwriting existing files without prompting.}
}
\description{
Generate a web page and javascript for a built-in example
}
