% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.R
\name{odin_js_versions}
\alias{odin_js_versions}
\title{Report JS versions}
\usage{
odin_js_versions()
}
\value{
A named list of \link{package_version} versions, for \code{odinjs}
and other coponents used in the JavaScript support.
}
\description{
Report versions of JavaScript packages used to run odin models.
}
