% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface-dev.R
\name{odin_validate_model}
\alias{odin_validate_model}
\title{Validate an odin model}
\usage{
odin_validate_model(x, type = NULL)
}
\arguments{
\item{x}{An expression, character vector or filename with the odin
code}

\item{type}{An optional string indicating the the type of input -
must be one of \code{expression}, \code{file} or \code{text} if
provided.  This skips the type detection code used by odin and
makes validating user input easier.}
}
\description{
Validate an odin model.  This function is closer to
\code{\link{odin_}} than \code{\link{odin}} because it does not do
any quoting of the code.  It is primarily intended for use within
other applications.
}
\author{
Rich FitzJohn
}
