% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.odin_model}
\alias{plot.odin_model}
\alias{plot.odin_generator}
\title{Plotting odin models}
\usage{
\method{plot}{odin_model}(x, ...)

\method{plot}{odin_generator}(x, ...)
}
\description{
The \code{plot} method for \code{odin_model} objects is a wrapper around the
\code{$code} slot of these objects.
}
