% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_bundle.R
\name{odin_js_bundle}
\alias{odin_js_bundle}
\title{Create a bundle of odin models}
\usage{
odin_js_bundle(filenames, dest = tempfile(), include = NULL,
  include_dopri = TRUE)
}
\arguments{
\item{filenames}{Filenames with odin source code}

\item{dest}{Destination file for the generated javascript}

\item{include}{Optional vector of paths of filenames to include
into the javascript bundle}

\item{include_dopri}{Logical, indicating if the dopri solver
should be included as well.}
}
\description{
Create a JavaScript bundle of odin models
}
